Allowed config file to be blank or partially filled. Config file now generated on first run (and removed from repo).
yaluConfig.py now checks for the config file's presence on run (and creates a
blank one if it is not found.

yaluConfig.example now contains the old example config file.

Blank configs are allowed because yaluConfig.py can now output a list of default
config options (in fvwm format) on demand. This is PipeRead on startup meaning
that the user's settings over-ride the defaults but if not over-ridden then the
system does not bug-out because the variable is undefined.